feat: Category 목록 CategoryFragment에 연결
