+++ app/src/main/res/layout/fragment_category.xml
@@ -59,12 +59,16 @@
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv_category"
-            android:layout_width="0dp"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginTop="20dp"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/et_history" />
+            app:layout_constraintTop_toBottomOf="@id/et_history"
+            app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
+            app:spanCount="3"
+            android:orientation="horizontal"
+            tools:listitem="@layout/item_category"/>
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
